pyproject.toml
[project]
name = "sentinel-code"
version = "0.1.0"
description = "AI-powered vulnerability scanner with threat intelligence mapping"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "httpx>=0.28.0",
    "google-genai>=1.0.0",
    "pydantic>=2.10.0",
    "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]
